Make common delete modal for components (#766)
