+++ dashboards-observability/.cypress/integration/7_app_analytics.spec.js
@@ -251,7 +251,7 @@ describe('Viewing application', () => {
   it('Adds filter when Trace group name is clicked', () => {
     cy.get('[data-test-subj="app-analytics-overviewTab"]').click();
     cy.get('[data-test-subj="dashboard-table-trace-group-name-button"]').contains('client_create_order').click();
-    cy.get('.euiTableRow').should('have.length', 1);
+    cy.get('.euiTableRow').should('have.length', 1, { timeout: timeoutDelay });
     cy.get('[data-test-subj="client_create_orderFilterBadge"]').should('exist');
     cy.get('[data-test-subj="filterBadge"]').click();
     cy.get('[data-test-subj="deleteFilterIcon"]').click();
@@ -586,7 +586,8 @@ describe('Application Analytics home page', () => {
     });
     cy.get('[data-test-subj="appAnalyticsActionsButton"]').click();
     cy.get('[data-test-subj="deleteApplicationContextMenuItem"]').click();
-    cy.get('[data-test-subj="confirmModalConfirmButton"]').click();
+    cy.get('[data-test-subj="popoverModal__deleteTextInput"]').type('delete');
+    cy.get('[data-test-subj="popoverModal__deleteButton"').click();
     cy.wait(delay);
     cy.get('.euiToast').contains(`Applications successfully deleted!`);
     cy.get(`[data-test-subj="${newName}ApplicationLink"]`).should('not.exist');
