+++ dashboards-observability/public/components/custom_panels/custom_panel_view.tsx
@@ -39,7 +39,7 @@ import {
 } from '../../../common/constants/custom_panels';
 import { SavedVisualizationType, VisualizationType } from '../../../common/types/custom_panels';
 import { PanelGrid } from './panel_modules/panel_grid';
-import { DeletePanelModal, getCustomModal } from './helpers/modal_containers';
+import { getCustomModal } from './helpers/modal_containers';
 import PPLService from '../../services/requests/ppl';
 import {
   isDateValid,
@@ -59,6 +59,7 @@ import {
   parseForIndices,
 } from '../common/search/autocomplete_logic';
 import { AddVisualizationPopover } from './helpers/add_visualization_popover';
+import { DeleteModal } from '../common/helpers/delete_modal';
 
 /*
  * "CustomPanelsView" module used to render an Operational Panel
@@ -224,7 +225,7 @@ export const CustomPanelView = (props: CustomPanelViewProps) => {
 
   const deletePanel = () => {
     setModalLayout(
-      <DeletePanelModal
+      <DeleteModal
         onConfirm={onDelete}
         onCancel={closeModal}
         title={`Delete ${openPanelName}`}
